4.0.0 Release (#301)
* migrated to challenge type agnostic plugin system
* renamed example plugin to indicate DNS type and created an HTTP equivalent
* updated plugin dev readme
* http plugins now use Domain,Token,Body params. Added WebRoot and WebSelfHost http plugins
* PluginArgs are now saved per-order and as JSON instead of XML. Import-PluginArgs has been replaced with (public) Get-PAPluginArgs. v3 PluginArgs will be auto-migrated for all orders when an account is selected.
* Updated GCloud and Azure plugins to better work with per-order plugin args
* Renamed Get-PAAuthorizations to Get-PAAuthorization
* change renewal window calc from hard 30 days to 1/3 cert lifetime with a cap of 30 days
* moved cert validation toggle into a private function
* updated Remove-* functions to use Import-PAConfig instead of setting script variables manually
* added a warning in Get-PAPluginArgs if specified order not found
* a whole bunch of test updates for Pester 5
* Added -AlwaysNewKey to New-PACertificate, New-PAOrder, and Set-PAOrder to enforce new key generation on renewal. Old per-operation -NewKey and -NewCertKey parameters have been removed. (#181)
* efficiency tweaks to Add-Member usage
* switched to more terse syntax for getting securestring plaintext
* replaced $script:OrderFolder with Get-OrderFolder for internal use
* made Update-PAOrder the only implementation for order saving
* Obfuscate PfxPass when serialized to disk (#207)
* Added PfxPassSecure parameter to New-PACertificate, New-PAOrder, and Set-PAOrder which takes precedence over PfxPass if specified. (#207)
* added 3072 as a common keylength for tab completion
* added buypass.com CA shortcuts for Set-PAServer
* added tab completion for Set-PAServer -DirectoryUrl
* Added -UseAltPluginEncryption to New-PAAccount and Set-PAAccount and -ResetAltPluginEncryption to Set-PAAccount to enable AES encryption for plugin parameters that works cross platform. (#150)
* added DnsPlugin aliases to Plugin params
* updates to Requirements and Platform support
* updated tutorial
* deprecate Invoke-HttpChallengeListener
* fix Set-CertValidation to only add ignore delegate if it doesn't already exist
* added telemetry ping to Submit-OrderFinalize. refactored Set-PAServer including ability to disable telemetry. added -Quiet to Get-PAServer
* all raw throw calls in public functions were converted to ThrowTerminatingException and Get-PAServer now warns when server doesn't exist instead of throwing
* overwrite backed up v3 config if it already exists
* ACME protocol HTTP details are no longer in verbose output. Debug output has been cleaned up a bit to make it easier to follow ACME related calls.
* error handling tweak in Submit-ChallengeValidation
* added DnsAlias and OCSPMustStaple params to Set-PAOrder
* Added Plugin, PluginArgs, DnsAlias, DnsSleep, and ValidationTimeout params to New-PAOrder and normalized capitalization of DnsSleep parameter across the module
* added Plugin column to default output view for PAOrder objects
* refactored New-PACertificate to account for param changes in New-PAOrder
* added ZEROSSL_PROD shortcut for Set-PASERVER
* refactored Submit-ChallengeValidation so that it only takes an Order object as a parameter. all of the previous validation parameters are read from the order object and can be modified via Set-PAOrder
* refactored Submit-OrderFinalize to work like Submit-ChallengeValidation where the active account is assumed and the order can be optionally passed via pipeline
* fixed Export-PACertFiles when the order has already expired
* Added Complete-PAOrder which performs the final steps on an order that has reached the 'ready' state
* added serial challenge validation support
* added basic SAN validation on signed cert
* added plugin dev guide tip for http challenges
* added a bit more error handling in Export-PACertFiles
* set all internal usage of ConvertTo-Json to -Depth 5 from the default of 2
* refactored internal plugin handling so validation happens during module load and simplifies use elsewhere
* add branch support to instdev.ps1
* added some verbose messaging to Publish/Unpublish-Challenge
* minor efficiency tweak in Set-PAOrder
* added ip identifier support
* allow P-521 EC keys to be used for accounts and certs
* added -NoRefresh switch to Set-PAServer
* Fixed Remove-PAServer not being able to remove a server that is unreachable and it no longer requires confirmation when there are no cached accounts.

+++ Posh-ACME/Private/Export-PACertFiles.ps1
@@ -7,15 +7,15 @@ function Export-PACertFiles {
     )
 
     # Make sure we have an account configured
-    if (!($acct = Get-PAAccount)) {
+    if (-not ($acct = Get-PAAccount)) {
         throw "No ACME account configured. Run Set-PAAccount or New-PAAccount first."
     }
 
     # Make sure we have an order
-    if (-not $Order -and !($Order = Get-PAOrder)) {
+    if (-not $Order -and -not ($Order = Get-PAOrder)) {
         throw "No ACME order specified and no current order selected. Run Set-PAOrder or specify an existing order object."
     }
-    $orderFolder = Join-Path $script:AcctFolder $Order.MainDomain.Replace('*','!')
+    $orderFolder = $Order | Get-OrderFolder
 
     # build output paths
     $certFile      = Join-Path $orderFolder 'cert.cer'
@@ -27,8 +27,10 @@ function Export-PACertFiles {
 
     if (-not $PfxOnly) {
 
-        # Re-download the cert/chains if the order has not expired.
+        # Download the cert+chain if the order has not expired.
         if ((Get-DateTimeOffsetNow) -lt [DateTimeOffset]::Parse($order.expires)) {
+
+            Write-Verbose "Downloading signed certificate"
 
             # build the header for the Post-As-Get request
             $header = @{
@@ -45,6 +47,30 @@ function Export-PACertFiles {
             } catch { throw }
 
             $pems = Split-PemChain -ChainBytes $response.Content
+
+            # Do some basic validation to make sure we got what we were expecting.
+            $cert = Import-Pem -InputString ($pems[0] -join "`n")
+            $altNames = $cert.GetSubjectAlternativeNames() | ForEach-Object {
+                if ($_[0] -eq [Org.BouncyCastle.Asn1.X509.GeneralName]::DnsName) {
+                    $_[1]
+                }
+                elseif ($_[0] -eq [Org.BouncyCastle.Asn1.X509.GeneralName]::IPAddress) {
+                    # gets returns as a hex string like "#01010101" that we need to parse
+                    ([ipaddress]([byte[]] -split ($_[1].Substring(1) -replace '..', '0x$& '))).ToString()
+                }
+            }
+            Write-Debug "SANs in downloaded cert: $(($altNames -join ', '))"
+            $orderNames = @($Order.MainDomain) + @($Order.SANs)
+            $orderNames | ForEach-Object {
+                if ($_ -notin $altNames) {
+                    Write-Error "$_ was requested but is not present in the list of Subject Alternative Names in the signed certificate."
+                }
+            }
+            $altNames | ForEach-Object {
+                if ($_ -notin $orderNames) {
+                    Write-Error "An extra name, $_, is present in the list of Subject Alternative Names in the signed certificate, but was not requested as part of the order."
+                }
+            }
 
             # write the lone cert
             Export-Pem $pems[0] $certFile
@@ -74,6 +100,7 @@ function Export-PACertFiles {
         }
         else {
             Write-Warning "Order has expired. Unable to re-download cert/chain files. Using cached copies."
+            $chain0File = Join-Path $orderFolder 'chain0.cer'
         }
 
         # try to find the chain file matching the preferred issuer if specified
@@ -86,12 +113,20 @@ function Export-PACertFiles {
             if (-not $selectedChainFile) {
                 Write-Warning "The preferred chain issuer, $($order.PreferredChain), was not found. Using the default chain."
                 $selectedChainFile = $chain0File
+            } else {
+                Write-Verbose "Using preferred chain issuer, $($order.PreferredChain)."
             }
         } else {
             $selectedChainFile = $chain0File
         }
 
         # build the appropriate chain and fullchain files
+        if (-not (Test-Path $certFile -PathType Leaf)) {
+            throw "Cert file not found: $certFile"
+        }
+        if (-not (Test-Path $selectedChainFile -PathType Leaf)) {
+            throw "Chain file not found: $selectedChainFile"
+        }
         Copy-Item $selectedChainFile $chainFile
         $fullchainLines = (Get-Content $certFile) + (Get-Content $selectedChainFile)
         Export-Pem $fullchainLines $fullchainFile
